+++ src/games/skyrimvr/skyrimvrsavegame.cpp
@@ -17,7 +17,7 @@ SkyrimVRSaveGame::SkyrimVRSaveGame(QString const& fileName, GameSkyrimVR const* 
 {
   FileWrapper file(fileName, u"TESV_SAVEGAME"_s);  // 10bytes
 
-  unsigned long version;
+  uint32_t version;
   FILETIME ftime;
   fetchInformationFields(file, version, m_PCName, m_PCLevel, m_PCLocation, m_SaveNumber,
                          ftime);
@@ -37,23 +37,23 @@ SkyrimVRSaveGame::SkyrimVRSaveGame(QString const& fileName, GameSkyrimVR const* 
   setCreationTime(dt);
 }
 
-void SkyrimVRSaveGame::fetchInformationFields(FileWrapper& file, unsigned long& version,
+void SkyrimVRSaveGame::fetchInformationFields(FileWrapper& file, uint32_t& version,
                                               QString& playerName,
-                                              unsigned short& playerLevel,
+                                              uint16_t& playerLevel,
                                               QString& playerLocation,
-                                              unsigned long& saveNumber,
+                                              uint32_t& saveNumber,
                                               FILETIME& creationTime) const
 {
-  unsigned long headerSize;
+  uint32_t headerSize;
   file.read(headerSize);  // header size "TESV_SAVEGAME"
   file.read(version);     // header version 74 (original Skyrim is 79)
   file.read(saveNumber);
 
   file.read(playerName);
 
-  unsigned long temp;
+  uint32_t temp;
   file.read(temp);
-  playerLevel = static_cast<unsigned short>(temp);
+  playerLevel = static_cast<uint16_t>(temp);
 
   file.read(playerLocation);
 
@@ -63,7 +63,7 @@ void SkyrimVRSaveGame::fetchInformationFields(FileWrapper& file, unsigned long& 
   QString race;
   file.read(race);  // race name (i.e. BretonRace)
 
-  file.skip<unsigned short>();  // Player gender (0 = male)
+  file.skip<uint16_t>();  // Player gender (0 = male)
   file.skip<float>(2);          // experience gathered, experience required
 
   file.read(creationTime);  // filetime
@@ -73,11 +73,11 @@ std::unique_ptr<GamebryoSaveGame::DataFields> SkyrimVRSaveGame::fetchDataFields(
 {
   FileWrapper file(getFilepath(), u"TESV_SAVEGAME"_s);  // 10bytes
 
-  unsigned long version = 0;
+  uint32_t version = 0;
   {
     QString dummyName, dummyLocation;
-    unsigned short dummyLevel;
-    unsigned long dummySaveNumber;
+    uint16_t dummyLevel;
+    uint32_t dummySaveNumber;
     FILETIME dummyTime;
 
     fetchInformationFields(file, version, dummyName, dummyLevel, dummyLocation,
@@ -86,8 +86,8 @@ std::unique_ptr<GamebryoSaveGame::DataFields> SkyrimVRSaveGame::fetchDataFields(
 
   std::unique_ptr<DataFields> fields = std::make_unique<DataFields>();
 
-  unsigned long width;
-  unsigned long height;
+  uint32_t width;
+  uint32_t height;
   file.read(width);
   file.read(height);
